# api/openapi-spec/swagger.yaml
basePath: /
definitions:
  dto.Article:
    properties:
      author:
        type: string
      content:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      language:
        type: string
      metadata:
        $ref: '#/definitions/dto.ArticleMetadata'
      subtitle:
        type: string
      title:
        type: string
      url:
        format: string
        type: string
    type: object
  dto.ArticleMetadata:
    properties:
      category:
        description: Content metadata
        type: string
      importedAt:
        description: System metadata
        type: string
      publishedAt:
        type: string
      sourceId:
        description: Essential source tracking
        type: string
      sourceName:
        type: string
    type: object
  dto.ArticleSearchResult:
    properties:
      article:
        allOf:
        - $ref: '#/definitions/dto.Article'
        description: Embedded Article struct for search results
      score:
        description: Score rank between 0 and 1
        type: number
      score_normalized:
        description: ScoreNormalized is the normalized(between 0-1) score
        type: number
    type: object
  dto.BooleanParams:
    properties:
      expression:
        type: string
      language:
        type: string
    type: object
  dto.MatchParams:
    properties:
      field:
        type: string
      fuzziness:
        type: string
      language:
        type: string
      operator:
        type: string
      query:
        type: string
    type: object
  dto.MultiMatchParams:
    properties:
      field_weights:
        additionalProperties:
          format: float64
          type: number
        type: object
      fields:
        items:
          type: string
        type: array
      language:
        type: string
      operator:
        type: string
      query:
        type: string
    type: object
  dto.PhraseParams:
    properties:
      fields:
        items:
          type: string
        type: array
      language:
        type: string
      query:
        type: string
      slop:
        type: integer
    type: object
  dto.QueryWrapper:
    properties:
      boolean:
        $ref: '#/definitions/dto.BooleanParams'
      match:
        $ref: '#/definitions/dto.MatchParams'
      multi_match:
        $ref: '#/definitions/dto.MultiMatchParams'
      phrase:
        $ref: '#/definitions/dto.PhraseParams'
    type: object
  dto.SearchRequest:
    properties:
      cursor:
        type: string
      query:
        $ref: '#/definitions/dto.QueryWrapper'
      size:
        type: integer
    type: object
  dto.SearchResponse:
    properties:
      has_more:
        type: boolean
      hits:
        items:
          $ref: '#/definitions/dto.ArticleSearchResult'
        type: array
      max_score:
        type: number
      next_cursor:
        type: string
      page_max_score:
        type: number
      total_matches:
        type: integer
    type: object
  dto.SemanticSearchResponse:
    properties:
      has_more:
        type: boolean
      hits:
        items:
          $ref: '#/definitions/dto.Article'
        type: array
      next_cursor:
        type: string
    type: object
info:
  contact:
    email: support@newshunter.com
    name: API Support
  description: A full-text search engine for exploring multilingual news headlines
    and articles
  license:
    name: Apache 2.0
    url: https://opensource.org/licenses/Apache-2.0
  termsOfService: http://swagger.io/terms/
  title: News Hunter API
  version: "1.0"
paths:
  /v1/articles/search:
    get:
      consumes:
      - application/json
      description: Simple text search with automatic field selection and weighting.
        Cacheable and bookmarkable. Application determines optimal search strategy
        based on index configuration.
      parameters:
      - description: SearchStringQuery query text
        example: '"climate change"'
        in: query
        name: q
        required: true
        type: string
      - description: 'Results per page (default: 100, max: 10000)'
        example: 10
        in: query
        name: size
        type: integer
      - description: Pagination cursor (base64-encoded from previous response)
        in: query
        name: cursor
        type: string
      - description: 'SearchStringQuery language: english, serbian (default: english)'
        example: '"english"'
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SearchStringQuery results with pagination metadata
          schema:
            $ref: '#/definitions/dto.SearchResponse'
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Simple query string search
      tags:
      - search
  /v1/articles/semantic_search:
    get:
      consumes:
      - application/json
      description: Perform vector-based semantic search using the embedding model.
        Supports pagination via cursor. Application determines optimal search strategy
        based on index configuration. Results are cacheable and bookmarkable.
      parameters:
      - description: SearchStringQuery query text
        example: '"climate change"'
        in: query
        name: q
        required: true
        type: string
      - description: 'Results per page (default: 100, max: 10000)'
        example: 10
        in: query
        name: size
        type: integer
      - description: Pagination cursor (base64-encoded from previous response)
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SearchStringQuery results with pagination metadata
          schema:
            $ref: '#/definitions/dto.SemanticSearchResponse'
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Vector-based semantic search
      tags:
      - search
  /v1/articles/structured:
    post:
      consumes:
      - application/json
      description: Execute structured search queries with explicit control over fields,
        weights, and operators. Supports match and multi_match query types. Follows
        Elasticsearch query DSL pattern.
      parameters:
      - description: Structured search request with query wrapper
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: SearchStringQuery results with pagination metadata
          schema:
            $ref: '#/definitions/dto.SearchResponse'
        "400":
          description: Bad request - invalid query structure or parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Query type not supported by searcher backend
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Structured search API
      tags:
      - search
swagger: "2.0"